Add shared executeTest function
