+++ client/api_mock.go
@@ -1,6 +1,8 @@
 package client
 
 import (
+	"encoding/json"
+	"fmt"
 	"github.com/stretchr/testify/require"
 	"io"
 	"net/http"
@@ -228,4 +230,29 @@ func FakeUAAServer(expiresIn int) *httptest.Server {
 func teardown() {
 	server.Close()
 	fakeUAAServer.Close()
+}
+
+func executeTests(tests []RouteTest, t *testing.T) {
+	for _, tt := range tests {
+		func() {
+			setup(tt.Route, t)
+			defer teardown()
+			details := fmt.Sprintf("%s %s", tt.Route.Method, tt.Route.Endpoint)
+			if tt.Description != "" {
+				details = tt.Description + ": " + details
+			}
+
+			c, _ := NewTokenConfig(server.URL, "foobar")
+			cl, err := New(c)
+			require.NoError(t, err, details)
+
+			obj, err := tt.Action(cl, t)
+			require.NoError(t, err, details)
+			if tt.Expected != "" {
+				actual, err := json.Marshal(obj)
+				require.NoError(t, err, details)
+				require.JSONEq(t, tt.Expected, string(actual), details)
+			}
+		}()
+	}
 }
